move the test to a unit test
